requirements.txt
black==22.3.0
boto3==1.24.4
botocore==1.27.4
click==8.1.3
jmespath==1.0.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
python-dateutil==2.8.2
python-dotenv==0.20.0
s3transfer==0.6.0
six==1.16.0
tomli==2.0.1
typing-extensions==4.2.0
urllib3==1.26.9
